requirements.txt
Flask
gunicorn
tensorflow
numpy
pillow
google-cloud-firestore
firebase-admin